Change listener from panel to calendarView
